Important classes that manage many bot functions.
Created the main class of the module tools, the Command class. This
manages user levels and cooldowns.

Created the Timer class, simply manages cooldowns. Has yet to be
tested.

Created the CommandModule class. This class manages all the Command
objects that are added to it when the module itself registers a new
command. It manages checks for duplicates, the addition and removal
of commands, as well as the parsing of a users message to see whether
or not the user was attempting to run one of its commands.

Created ModerationModule class. This class is incredibly simple at
its current point and just checks all of its filters to see if any
are triggered by the users message.

Created GeneralCommandDatabase class. This class contains all the
tools necessary to manage a simple database of custom commands. It
Allows for the addition, removal, duplicate checks, cooldowns, user
levels, and getting the response of the command.