Basic tools for managing modules
Created storage system, including a function for allowing other
modules to request their storage directory. Directories are only
created on request.

Created a basic user level checker for use by default on creation
of commands.

Created a function for checking for duplicates.

Created a function for reserving command names, to prevent issues
with custom (general) commands conflicting.

Created function for registering a command module, allowing a module
to register commands.

Created a function for registering commands.

Created a function for registering shutdown functions, these are
meant to be called when the bot is shutting down.